requirements.txt
pyserial
systemd-python

